refactor: Command 위치 변경
