+++ src/main/java/chess/controller/Controller.java
@@ -2,6 +2,7 @@ package chess.controller;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.game.ChessGame;
+import chess.domain.game.Command;
 import chess.domain.game.Result;
 import chess.domain.piece.Color;
 import chess.view.InputView;
